add package signing on publish (#110)
* add package signing on publish

* breaking change: tag is a separate picee

* add public key importing

* more key management for installing/publishing

* implement much more of the core installing logic

* include omre info in failures/handle more cases

* fix som ebugs, improve error messages

* refactor code involving SQL

* rm tx from migration file

* rename public_key_tag -> public_key_id

* store private key as canonical path

* Add tests

* rm dbg

* apply many changes to intergrate with server

* large number of fixes/updates + regression test

* rm unused test

* update from feedback

+++ src/commands/keys.rs
@@ -1,8 +1,11 @@
 //! Subcommand to deal with keys for signing wapm packages
 
+use crate::graphql::{self, DateTime};
 use crate::keys::*;
+use crate::util;
+
+use graphql_client::*;
 use prettytable::{format, Table};
-use std::io::Write;
 use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
@@ -12,8 +15,12 @@ pub enum KeyOpt {
     List(List),
 
     #[structopt(name = "register")]
-    /// Register a key with wapm
+    /// Register a personal key pair with wapm
     Register(Register),
+
+    #[structopt(name = "import")]
+    /// Import a public key from somewhere
+    Import(Import),
 
     #[structopt(name = "delete")]
     /// Delete a keypair from wapm
@@ -40,11 +47,27 @@ pub struct Register {
     private_key_location: String,
 }
 
+#[derive(GraphQLQuery)]
+#[graphql(
+    schema_path = "graphql/schema.graphql",
+    query_path = "graphql/queries/publish_public_key.graphql",
+    response_derives = "Debug"
+)]
+struct PublishPublicKeyMutation;
+
 /// Deletes a key to wapm
 #[derive(StructOpt, Debug)]
 pub struct Delete {
     /// The identifier of the public key
-    public_key: String,
+    public_key_id: String,
+}
+
+/// Import a public key from somewhere else
+#[derive(StructOpt, Debug)]
+pub struct Import {
+    #[structopt(long = "user-name")]
+    user_name: String,
+    public_key_value: String,
 }
 
 pub fn keys(options: KeyOpt) -> Result<(), failure::Error> {
@@ -83,33 +106,58 @@ pub fn keys(options: KeyOpt) -> Result<(), failure::Error> {
             public_key_location,
             private_key_location,
         }) => {
-            // mutate server
-            add_personal_key_pair_to_database(
+            let (pk_id, pk_v, tx) = add_personal_key_pair_to_database(
                 &mut key_db,
                 public_key_location.clone(),
                 private_key_location.clone(),
             )?;
-            println!("Key pair successfully added!")
+            let q = PublishPublicKeyMutation::build_query(publish_public_key_mutation::Variables {
+                key_id: pk_id.clone(),
+                key: pk_v,
+                verifying_signature_id: None,
+            });
+            let response_or_err: Result<publish_public_key_mutation::ResponseData, _> =
+                graphql::execute_query(&q);
+            match response_or_err {
+                Ok(_) => {
+                    tx.commit().map_err(|e| {
+                        format_err!(
+                            "Failed to store key pair in local database: {}",
+                            e.to_string()
+                        )
+                    })?;
+                    println!("Key pair successfully added!")
+                }
+                Err(e) => {
+                    error!("Failed to upload public key to server: {}", e);
+                    #[cfg(feature = "telemetry")]
+                    sentry::integrations::failure::capture_error(&e);
+                }
+            };
         }
-        KeyOpt::Delete(Delete { public_key }) => {
+        KeyOpt::Delete(Delete { public_key_id }) => {
             let full_public_key =
-                get_full_personal_public_key_by_pattern(&key_db, public_key.clone())?;
+                get_full_personal_public_key_by_id(&key_db, public_key_id.clone())?;
             warn!(
                 "You are about to delete the key pair associated with {:?} from wapm.\nThis cannot be undone.",
                 &full_public_key
             );
-            print!("Please confirm that you want to permanently delete this key pair from wapm:\n[y/n] ");
-            std::io::stdout().flush()?;
-            let mut input_str = String::new();
-            std::io::stdin().read_line(&mut input_str)?;
-            match input_str.to_lowercase().trim_end() {
-                "yes" | "y" => {
-                    delete_key_pair(&mut key_db, full_public_key)?;
-                }
-                _ => {
-                    println!("Aborting");
-                }
+            let user_confirmed_key_deletion = util::prompt_user_for_yes(
+                "Please confirm that you want to permanently delete this key pair from wapm:",
+            )?;
+            if user_confirmed_key_deletion {
+                delete_key_pair(&mut key_db, full_public_key)?;
+            } else {
+                println!("Aborting");
             }
+        }
+        KeyOpt::Import(Import {
+            user_name,
+            public_key_value,
+        }) => {
+            let user_name = user_name.trim().to_string();
+            let (pk_id, pkv) = normalize_public_key(public_key_value)?;
+            import_public_key(&mut key_db, &pk_id, &pkv, user_name)?;
         }
     }
 
@@ -119,11 +167,18 @@ pub fn keys(options: KeyOpt) -> Result<(), failure::Error> {
 pub fn create_personal_key_table(keys: Vec<PersonalKey>) -> Result<String, failure::Error> {
     let mut table = Table::new();
     table.set_format(*format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
-    table.add_row(row!["KEY", "ACTIVE", "PRIVATE KEY LOCATION", "DATE ADDED"]);
+    table.add_row(row![
+        "TAG",
+        "ACTIVE",
+        "KEY",
+        "PRIVATE KEY LOCATION",
+        "DATE ADDED"
+    ]);
     for key in keys {
         table.add_row(row![
-            key.public_key_value,
+            key.public_key_id,
             key.active,
+            key.public_key_value,
             key.private_key_location.unwrap_or("None".to_string()),
             time::strftime("%Y-%m-%d", &time::at(key.date_created))?
         ]);
@@ -134,10 +189,11 @@ pub fn create_personal_key_table(keys: Vec<PersonalKey>) -> Result<String, failu
 pub fn create_wapm_public_key_table(keys: Vec<WapmPublicKey>) -> Result<String, failure::Error> {
     let mut table = Table::new();
     table.set_format(*format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
-    table.add_row(row!["USER", "KEY", "DATE ADDED"]);
+    table.add_row(row!["USER", "TAG", "KEY", "DATE ADDED"]);
     for key in keys {
         table.add_row(row![
             key.user_name,
+            key.public_key_id,
             key.public_key_value,
             time::strftime("%Y-%m-%d", &time::at(key.date_created))?
         ]);
